:construction: converting from jobs to commands

+++ commands/maintenance.php
@@ -1,36 +1,48 @@
 <?php
 
+use Bnomei\Janitor;
 use Kirby\CLI\CLI;
 
 return [
     'description' => 'Maintenance',
     'args' => [
         'up' => [
-            'prefix'      => 'u',
             'longPrefix'  => 'up',
             'description' => 'Up',
             'noValue'     => true,
         ],
         'down' => [
-            'prefix'      => 'd',
             'longPrefix'  => 'down',
             'description' => 'Down',
             'noValue'     => true,
         ],
-    ],
+    ] + Janitor::ARGS,
     'command' => static function (CLI $cli): void {
+        $message = '';
         $maintenance = kirby()->roots()->index() . '/.maintenance';
+        $down = !file_exists($maintenance); // toggle
         if ($cli->arg('up')) {
-            if (file_exists($maintenance)) {
-                unlink($maintenance);
-            }
-            $cli->success('UP');
+            $down = false;
         }
         if ($cli->arg('down')) {
-            file_put_contents(kirby()->roots()->index() . '/.maintenance', (string) time());
-            $cli->success('DOWN');
+            $down = true;
         }
 
-        $cli->error('Missing argument `u`/`up` or `d`/`down`.');
+        if ($down === false) {
+            file_exists($maintenance) && unlink($maintenance);
+            $message = 'UP ' . date('c') . ' [' . $cli->arg('user') . ']';
+            defined('STDOUT') && $cli->success($message);
+        }
+        elseif ($down === true) {
+            $message = 'DOWN ' . date('c') . ' [' . $cli->arg('user') . ']';
+            file_put_contents($maintenance, $message);
+            defined('STDOUT') && $cli->red($message);
+        }
+
+        janitor()->data($cli->arg('command'), [
+            'status' => $down ? 203 : 200,
+            // urls forwarded to janitor in `download` will trigger a download in panel.
+            'label' => $message,
+        ]);
     }
 ];
